Moved request structs to namespace files
Request structures were stored on separate files per API method. Moved
all structs to be grouped on a separate files per API call namespace.
E.g. `user.something` structs are stored on `user.go` and so on.
