--- requests/differential_getcommitpaths.go
@@ -1,8 +1,0 @@
-package requests
-
-// DifferentialGetCommitPathsRequest represents a request to the
-// differential.getcommitpaths call.
-type DifferentialGetCommitPathsRequest struct {
-	RevisionID uint64 `json:"revision_id"`
-	Request
-}
