+++ requests/diffusion.go
