--- requests/maniphest_gettasktransactions.go
@@ -1,7 +1,0 @@
-package requests
-
-// ManiphestGetTaskTransactions represents a request to maniphest.gettasktransactions.
-type ManiphestGetTaskTransactions struct {
-	IDs []string `json:"ids"`
-	Request
-}
