--- requests/phid_lookup.go
@@ -1,6 +1,0 @@
-package requests
-
-type PHIDLookupRequest struct {
-	Names []string `json:"names"`
-	Request
-}
