--- requests/phid_query.go
@@ -1,6 +1,0 @@
-package requests
-
-type PHIDQueryRequest struct {
-	PHIDs []string `json:"phids"`
-	Request
-}
